+++ extensions/youtube/src/main/java/app/morphe/extension/youtube/patches/DisablePlayerPopupPanelsPatch.java
@@ -7,7 +7,10 @@ public class DisablePlayerPopupPanelsPatch {
     /**
      * Injection point.
      */
-    public static boolean disablePlayerPopupPanels() {
-        return Settings.DISABLE_PLAYER_POPUP_PANELS.get();
+    public static boolean disablePlayerPopupPanels(boolean value1, boolean value2) {
+        if (!Settings.DISABLE_PLAYER_POPUP_PANELS.get()) {
+            return false;
+        }
+        return value1 && value2;
     }
 }